Use UIDs instead of CS for PatientStudyIDs
- Rename PatientStudyIDs to PatientStudyInstanceUIDs; the latter
name is longer but more precise; it may be a pain to change this now,
but it seems like it will become more painful later, so is worth
addressing now
- Make PatientStudyInstanceUIDs use the `UID` VR type, insted of `CS`;
given that the StudyInstanceUIDs that are used to build up PatientStudyInstanceUIDs
are UIDs, this seems appropriate.  Unfortuneatly, the UID type (oddly)
warns when you use equality checks.  In particular, the warning is:

```
pycom/uid.py:73: DeprecationWarning: The equality test for "UID == 'nonexistentseriesID'" is deprecated and will be removed in pydicom v1.2. In the future use "UID.name == 'nonexistentseriesID'"
```

This warning is still being thrown in other parts of the code base, so
we should address this next time we touch these parts.