Set missing "additional_tags" to blank strings
It seems like DICOM tends to stick in blank strings when information is
missing.  Thus, it seems that DICOM, like javascript, distinguishes
between `null` and `undefined`.  In DICOM a blank string is equivalent
to `null` and a missing attribute is equivalent to `undefined`.

When a user queries for additional tags, it seems appropriate to stick
`null` in the fields that were not present in the datasets.

We should discuss and possibly apply this behaviour to all of the other
methods too.